+++ cilantro/utils/__init__.py
@@ -1,2 +1,3 @@
 from cilantro.utils.utils import Encoder, TestNetURLHelper
 from cilantro.utils.picklable_mixin import PicklableMixin
+from cilantro.utils.lprocess import LProcess
