+++ playground/error_logging/test2.py
@@ -1,32 +1,12 @@
 from cilantro.logger import get_logger
-from multiprocessing import Process
+from cilantro.utils import LProcess
 import asyncio
 import time
-import traceback
-import inspect
 
-DELIM_LEN = 60
-OUTER_DELIM = '!'
-INNER_DELIM = '-'
 
-class LProcess(Process):
-    def run(self):
-        log = get_logger(self.name)
-        log.info("---> {} Starting --->".format(self.name))
 
-        try:
-            super().run()
-        except Exception as e:
-            err_msg = '\n' + OUTER_DELIM * DELIM_LEN
-            err_msg += '\nException caught on ' + self.name + ':\n' + str(e)
-            err_msg += '\n' + INNER_DELIM * DELIM_LEN
-            err_msg += '\n' + traceback.format_exc()
-            err_msg += '\n' + INNER_DELIM * DELIM_LEN
-            err_msg += '\n' + OUTER_DELIM * DELIM_LEN
-            log.error(err_msg)
-        finally:
-            log.info("<--- {} Terminating <---".format(self.name))
-            # TODO -- signal to parent to call .join() on this process to clean up nicely
+
+
 
 
 def sub_sub_proc_start():
@@ -67,8 +47,8 @@ def proc_start():
 
     # p.join()
 
-    # log.critical("!!!!\n!!!!!\nABOUT TO EXPLODE\n!!!!!\n!!!!!")
-    # i = 10 / 0
+    log.critical("!!!!\n!!!!!\nABOUT TO EXPLODE\n!!!!!\n!!!!!")
+    i = 10 / 0
